+++ run.py
@@ -13,7 +13,7 @@ def main(config: DictConfig):
     # Imports can be nested inside @hydra.main to optimize tab completion
     # https://github.com/facebookresearch/hydra/issues/934
     from src.train import train
-    from src.utils import utils
+    from src.utils import lht_utils
 
     OmegaConf.resolve(config)
     # A couple of optional utilities:
@@ -21,11 +21,11 @@ def main(config: DictConfig):
     # - forcing debug-friendly configuration
     # - verifying experiment name is set when running in experiment mode
     # You can safely get rid of this line if you don't want those
-    utils.extras(config)
+    lht_utils.extras(config)
 
     # Pretty print config using Rich library
     if config.get("print_config"):
-        utils.print_config(config, resolve=True)
+        lht_utils.print_config(config, resolve=True)
 
     # Train model
     return train(config)